Refactor file handler into modules
